apollo: update libmmcamera_faceproc.so from bacon
Source:
https://github.com/TheMuppets/proprietary_vendor_oneplus/blob/cm-14.1/bacon/proprietary/vendor/lib/libmmcamera_faceproc.so

In android 6.0 there was a warning when loading libmmcamera_faceproc.so:
libmmcamera_faceproc.so has text relocations. This is wasting memory and
is a security risk. Please fix.

With the upgrade to android 7.0 linker now returns an error instead and
this lib fails to load.
By updating to bacon's latest binary from a year ago it fixes camera view.